# Wait-free `SignalState` reads. Optimization for one-observer `Signal`s.