1. Grabbing the observers is now wait-free, at the cost of insertion and removal of observers being more expensive.
2. Optimization for one-observer `Signal`s.

#### Implementation note
Unless user explicitly synchronises the input and the removal of an observer, removing an observer would not free the observer from receiving events in a concurrent environment _immediately_, but only _eventually_. This has been the case since 3.x, since concurrent senders all grab a copy of the observer bag.

[Benchmark Source (A ported test from CwlSignal)](https://gist.github.com/andersio/70ded6fb313df4ebeed19e98b50ea77a)

`master`
```
AsyncMap
Performance is 9.032114933 seconds versus expected 3.3. Rate is 110716.040198555 per second.

SingleSignal
Performance is 8.795302343 seconds versus expected 3.0. Rate is 1136970.57929553 per second.
```

Read Copy Update
```
AsyncMap
Performance is 6.655907553 seconds versus expected 3.3. Rate is 150242.471374061 per second.
SingleSignal
Performance is 5.766949619 seconds versus expected 3.0. Rate is 1734018.96334479 per second.
```

Read Copy Update + `SignalObservers`
```
AsyncMap
Performance is 5.134279781 seconds versus expected 3.3. Rate is 194769.284623058 per second.
SingleSignal
Performance is 3.648229742 seconds versus expected 3.0. Rate is 2741055.44529602 per second.
```