@andersio 2016-12-03
Blocked by #137.